NMEA parser removed from the basic mBed-OS wxample
Removed the displayInfo function as that would also parse the the raw NMEA sentences, which was not intended for the most basic example. That would confuse people as this mBed port does not work with the original version of TinyGPS++ library.

Would someone want to parse NMEA sentences, they will need to add codygray's port of the TinyGPS++ library to the mBed example code.